+++ ib1/ib1.psm1
@@ -8,11 +8,46 @@ $driverFolder='C:\Dell'
 $logFile="$env:TEMP\ib1.log"
 $skillpipeUrl='https://prod-sp-ereader-assets.azureedge.net/WPFReader/skillpipeReaderSetup.exe'
 $mslearnGitUrl='https://api.github.com/repos/MicrosoftLearning/'
+$mslearnGit='MicrosoftLearning'
 $defaultSwitchId='c08cb7b8-9b3c-408e-8e30-5e16a3aeb444'
 $logStart=$true
 $repoParam=@{
-  'M10979'=@{'repo'='10979-MicrosoftAzureFundamentals';'srcPath'='AllFiles';'destPath'='E:\';'shortcut'='MicrosoftLearning/10979-MicrosoftAzureFundamentals/tree/master/Instructions';'vmName'='mia-cl1'}
-  'm20533'=@{'repo'='20533-ImplementingMicrosoftAzureInfrastructureSolutions';'srcPath'='Allfiles';'destPath'='F:\';'shortcut'='MicrosoftLearning/20533-ImplementingMicrosoftAzureInfrastructureSolutions/tree/master/Instructions';'vmName'='mia-cl1'}}
+  'm10979'='
+  new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/10979-MicrosoftAzureFundamentals/tree/master/Instructions" -title "Ateliers stage m10979";
+  if ($env:COMPUTERNAME -like "pc-formateur") {get-ib1Repo 10979-MicrosoftAzureFundamentals -srcPath Allfiles -destPath E:\}';
+  'm20533'='
+  new-ib1Shortcut -URL "https://github.com/MicrosoftLearning/20533-ImplementingMicrosoftAzureInfrastructureSolutions/tree/master/Instructions" -title "Ateliers stage m20533";
+  if ($env:COMPUTERNAME -like "pc-formateur") {get-ib1Repo 20533-ImplementingMicrosoftAzureInfrastructureSolutions -srcPath Allfiles -destPath F:\}';
+  'msaz100'='
+  get-ib1Repo AZ-100-MicrosoftAzureInfrastructureDeployment
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-100-MicrosoftAzureInfrastructureDeployment\"
+  rename-item -path $dest -newName "Ateliers MSAZ100" -errorAction SilentlyContinue
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\Ateliers MSAZ100\"
+  remove-item "$($dest)AZ-100T03A-ENU-LabFiles.zip" -force -errorAction SilentlyContinue
+  remove-item "$($dest)AZ-100T04A-ENU-LabFiles.zip" -force -errorAction SilentlyContinue
+  Add-Type -AssemblyName System.IO.Compression.FileSystem
+  get-childitem ($dest)|foreach-object {unzip $_.fullName $dest;remove-item $_.fullName -force -errorAction SilentlyContinue}
+  get-childitem ($dest) -directory|foreach-object {move-item "$($_.fullname)\*" -destination $dest;remove-item $_.fullName -force}
+  get-childitem ($dest) -file|foreach-object {rename-item -path $_.fullName -newName "Partie $($_.name[8]).pdf"}
+  invoke-webRequest -uri https://raw.githubusercontent.com/renaudwangler/ib/Master/ib1/extra/AZ-100IntroA.pptx -OutFile "$env:userprofile\documents\AZ-100IntroA.pptx"';
+  'msaz101'='
+  get-ib1Repo AZ-101-MicrosoftAzureIntegrationandSecurity;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\AZ-101-MicrosoftAzureIntegrationandSecurity\";
+  rename-item -path $dest -newName "Ateliers MSAZ101" -errorAction SilentlyContinue;
+  $dest=[Environment]::GetFolderPath("CommonDesktopDirectory")+"\Ateliers MSAZ101\";
+  remove-item "$($dest)AZ-101T03A-ENU-LabFiles.zip" -force;
+  remove-item "$($dest)AZ-101T04A-ENU-LabFiles.zip" -force;
+  Add-Type -AssemblyName System.IO.Compression.FileSystem
+  get-childitem ($dest)|foreach-object {unzip $_.fullName $dest;remove-item $_.fullName -force -errorAction SilentlyContinue}
+  get-childitem ($dest) -directory|foreach-object {move-item "$($_.fullname)\*" -destination $dest;remove-item $_.fullName -force}
+  get-childitem ($dest) -file|foreach-object {rename-item -path $_.fullName -newName "Partie $($_.name[8]).pdf"}
+  invoke-webRequest -uri https://raw.githubusercontent.com/renaudwangler/ib/Master/ib1/extra/AZ-101IntroA.pptx -OutFile "$env:userprofile\documents\AZ-101IntroA.pptx"'}
+
+
+function Unzip {
+param([string]$zipfile,[string]$outpath)
+  [System.IO.Compression.ZipFile]::ExtractToDirectory($zipfile, $outpath)
+}
 
 function write-ib1log {
 [CmdletBinding(DefaultParameterSetName='TextLog')]
@@ -178,64 +213,72 @@ param([string]$VMName,[bool]$exactVMName=$false)
 function get-ib1Repo {
 <#
 .SYNOPSIS
-Cette commande permet de récupérer en local le contenu d'un repositery Git.
+Cette commande permet de copier en local le contenu d'un repositery Git.
 .PARAMETER Repo
-Nom du repositery sur GitHub
+Nom du repositery sur GitHub (sera préfixé de l'indentifiant MS learning en l'absence de caractère "/")
 .PARAMETER srcPath
-Chemin des fichiers à récupérer dans le Repositery
+Chemin des fichiers à récupérer dans le Repositery (par défaut "/")
 .PARAMETER destPath
-Chemin local ou seront posés les fichier récupérés du Git.
+Chemin local ou seront posés les fichier récupérés du Git (par défaut, sur le bureau de l'utilisateur actuel)
 .PARAMETER course
 Référence du stage pour paramètrage simplifié (remplaçe les 3 paramètres précédents).
-Nota : Si utilisé dans la VM, le paramètre "course" récupère les fichiers des labs,
-       Si utilisé hors VM, le paramètre "course" crée un raccourci vers les instructions des ateliers.
+Valeurs possibles : m10979, m20533, msaz100 et msAZ101
 .EXAMPLE
 get-ib1Repo -repo '10979-MicrosoftAzureFundamentals' -srcPath 'Allfiles' -destPath 'c:\10979'
 Récupère les fichiers contenu dans le répertoire 'AllFiles' du repo '10979' et le copie dans le repertoire local c:\10979
 .EXAMPLE
-get-ib1Repo -course m10979
-Récupère les fichiers pour le stage m10979
+get-ib1Repo -course msAZ100
+Récupère les fichiers pour le stage msAZ100
 #>
 [CmdletBinding(
-DefaultParameterSetName='VMName')]
+DefaultParameterSetName='Repo')]
 PARAM(
 [string]$Repo,
-[string]$srcPath,
-[string]$destPath='E:',
+[string]$srcPath='*',
+[string]$destPath='',
 [string]$course='')
 begin{get-ib1elevated $true; compare-ib1PSVersion "4.0"}
 process {
 if ($course -ne '') {
   if (-not $repoParam.$course) {write-ib1log "Le paramètre simplifié -course ne peut avoir que l'une des valeurs suivantes: $($repoParam.Keys). Merci de vérifier!" -ErrorLog}
-  if ($env:COMPUTERNAME -notlike $repoParam.$course.vmName) {
-    write-ib1log "Création du raccourci pour les instructions d'atelier '$course' sur le bureau."
-    new-ib1Shortcut -URL "https://github.com/$($repoParam.$course.shortcut)" -title "instructions d'atelier - $course"
-    break}
-  $repo=$repoParam.$course.repo
-  $srcPath=$repoParam.$course.srcPath
-  $destpath=$repoParam.$course.destPath}
-if (-not $repo -or $repo -eq '') {write-ib1log "Le paramétre -Repo est manquant, merci de vérifier !" -ErrorLog}
-if (-not $srcPath -or $srcPath -eq '') {write-ib1log "Le paramétre -srcPath est manquant, merci de vérifier !" -ErrorLog}
+  else {
+    write-ib1log "Mise en place de l'environnement de stage pour le stage '$course'."
+    Invoke-Expression $repoParam.$course
+    break}}
+if (-not $Repo -or $Repo -eq '') {write-ib1log "Le paramétre -Repo est manquant, merci de vérifier !" -ErrorLog}
+if (-not $Repo.Contains('/')) {$repo=$mslearnGit+'/'+$Repo}
+#création du nom du répertoire local si non spécifié
+if ($destPath -eq '') {$destPath=[Environment]::GetFolderPath("CommonDesktopDirectory")+'\'+$Repo.substring($Repo.IndexOf('/')+1)}
+$ErrorActionPreference='silentlyContinue'
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-$srcFolder=$mslearnGitUrl+$repo+'/zipball'
-$destZip=$env:TEMP+'\'+$repo+'.zip'
+$gitRequest=[System.Net.WebRequest]::Create("https://github.com/$Repo")
+$gitResponse=$gitRequest.GetResponse()
+if ([int]$gitResponse.StatusCode -ne 200) {write-ib1log "L'url 'https://github.com/$Repo' n'existe pas ou est injoignable." -ErrorLog}
+$ErrorActionPreference='continue'
+if (-not (Test-Path $destPath)) {
+  if (-not (Test-Path $destPath.substring(0,$destPath.LastIndexOf('\')))) {write-ib1log "Le répertoire '$destPath' n'existe pas et ne peut être créé." -ErrorLog}
+  else {
+    write-ib1log "Le répertoire '$destPath' n'existe pas et sera créé." -DebugLog
+    New-Item $destPath -ItemType Directory|Out-Null}}
+$srcFolder="https://api.github.com/repos/$Repo/zipball"
+$destZip=$env:TEMP+'\'+$Repo.substring($Repo.IndexOf('/')+1)+'.zip'
 write-ib1log "Récupération du fichier '$srcFolder' dans '$destZip'." -DebugLog
 Invoke-WebRequest -Uri $srcFolder -OutFile $destZip
 Add-Type -AssemblyName System.IO.Compression.FileSystem
-write-ib1log "Création du dossier '$destPath'." -DebugLog
-if ($destPath.Length -gt 3) {New-Item -Path $destPath -ItemType Directory -Force|Out-Null}
+if ($srcPath -eq '*') {$srcFind='*'} else {$srcFind="*/$srcPath*"}
 $zip=[IO.Compression.ZipFile]::OpenRead($destZip)
-$zip.Entries|Where-Object {$_.FullName -like "*$repo*/$srcPath/*"}|ForEach-Object {
-  $destName=$_.FullName -replace "(?:[^\/]*)\/$srcPath"
+$zip.Entries|Where-Object {$_.FullName -like $srcFind}|ForEach-Object {
+  if ($srcPath -ne '*') {$destName=$_.FullName -replace "(?:[^\/]*)\/$srcPath"} else {$destName=$_.FullName.substring($_.FullName.IndexOf('/'))}
   if ($_.Name -eq '') {
     write-ib1log -progressTitleLog "Décompression des fichiers." "Création du dossier '$destName'."
     New-Item -Path $destPath$destName -ItemType Directory -Force -ErrorAction SilentlyContinue |Out-Null}
-  else {
+  elseif ($_.Name -notlike '*readme.md') {
     write-ib1log -progressTitleLog "Décompression des fichiers." "Décompression du fichier '$($_.Name)'."
-    if(![System.IO.File]::Exists($destPath+$destName)) {[IO.Compression.ZipFileExtensions]::ExtractToFile($_,$destPath+$destName)}}}
+    if(![System.IO.File]::Exists($destPath+$destName)) {[IO.Compression.ZipFileExtensions]::ExtractToFile($_,$destPath+$destName)|Out-Null}}}
 $zip.dispose()
 write-ib1log -progressTitleLog "Décompression des fichiers."
 Remove-Item -Path $destZip -Force}}
+
 
 function set-ib1VMCheckpointType {
 <#
@@ -1071,17 +1114,14 @@ param(
 [string]$GatewaySubnet='172.16.0.0',
 [string]$GatewayMask=24)
 get-ib1elevated $true
-compare-ib1PSVersion "4.0"
+compare-ib1PSVersion "5.0"
 write-ib1log -progressTitleLog "Mise en place des otpion nécessaires pour WinRM" "Passage des réseaux en privé."
 Get-NetConnectionProfile|where {$_.NetworkCategory -notlike '*Domain*'}|Set-NetConnectionProfile -NetworkCategory Private
-#write-ib1log -progressTitleLog "Mise en place des otpion nécessaires pour WinRM" "Option de confiance pour toutes machines."
-#Set-Item WSMan:\localhost\Client\TrustedHosts -value * -Force
 Set-ItemProperty –Path HKLM:\System\CurrentControlSet\Control\Lsa –Name ForceGuest –Value 0 -Force
 write-ib1log -progressTitleLog "Mise en place des otpion nécessaires pour WinRM" "Activation de PSRemoting."
 Enable-PSRemoting -Force|Add-Content -Path $logFile -Encoding UTF8
 write-ib1log -progressTitleLog "Mise en place des otpion nécessaires pour WinRM"
 if ((get-WindowsOptionalFeature -FeatureName Microsoft-Hyper-V-All -Online).state -eq 'enabled') {
-  new-ib1Nat
   if ((get-VMHost).EnableEnhancedSessionMode) {
     write-ib1log -progressTitleLog "Paramètrage de Hyper-V" "Désactivation de la stratégie de session avançée d'Hyper-V."
     Set-VMHost -EnableEnhancedSessionMode $false}
@@ -1286,5 +1326,6 @@ else {
 Set-Alias ibReset reset-ib1VM
 Set-Alias set-ib1VhdBoot mount-ib1VhdBoot
 Set-Alias complete-ib1Setup complete-ib1Install
+Set-Alias get-ib1Git get-ib1repo
 Export-moduleMember -Function install-ib1Chrome,complete-ib1Install,invoke-ib1NetCommand,new-ib1Shortcut,Reset-ib1VM,Mount-ib1VhdBoot,Remove-ib1VhdBoot,Switch-ib1VMFr,Test-ib1VMNet,Connect-ib1VMNet,Set-ib1TSSecondScreen,Import-ib1TrustedCertificate, Set-ib1VMCheckpointType, Copy-ib1VM, repair-ib1VMNetwork, start-ib1SavedVMs, get-ib1log, get-ib1version, stop-ib1ClassRoom, new-ib1Nat, invoke-ib1Clean, invoke-ib1Rearm, get-ib1Repo, set-ib1VMExternalMac
 Export-ModuleMember -Alias set-ib1VhdBoot,ibreset,complete-ib1Setup